associated orgId with all elements

+++ backend/lib/actions/lab.js
@@ -1,5 +1,4 @@
 "use strict";
-"use server";
 var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
     var desc = Object.getOwnPropertyDescriptor(m, k);
@@ -16,23 +15,13 @@ var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (
 }) : function(o, v) {
     o["default"] = v;
 });
-var __importStar = (this && this.__importStar) || (function () {
-    var ownKeys = function(o) {
-        ownKeys = Object.getOwnPropertyNames || function (o) {
-            var ar = [];
-            for (var k in o) if (Object.prototype.hasOwnProperty.call(o, k)) ar[ar.length] = k;
-            return ar;
-        };
-        return ownKeys(o);
-    };
-    return function (mod) {
-        if (mod && mod.__esModule) return mod;
-        var result = {};
-        if (mod != null) for (var k = ownKeys(mod), i = 0; i < k.length; i++) if (k[i] !== "default") __createBinding(result, mod, k[i]);
-        __setModuleDefault(result, mod);
-        return result;
-    };
-})();
+var __importStar = (this && this.__importStar) || function (mod) {
+    if (mod && mod.__esModule) return mod;
+    var result = {};
+    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
+    __setModuleDefault(result, mod);
+    return result;
+};
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
     function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
@@ -56,13 +45,19 @@ function createLab(JWTtoken, name, semester, batches, teachers, rooms, timetable
     return __awaiter(this, void 0, void 0, function* () {
         try {
             const { status, user } = yield auth.getPosition(JWTtoken);
+            if ((user === null || user === void 0 ? void 0 : user.orgId) == null) {
+                return {
+                    status: statusCodes_1.statusCodes.BAD_REQUEST,
+                    data: null,
+                };
+            }
             if (status == statusCodes_1.statusCodes.OK && user && user.role != "viewer") {
                 const existingLab = yield prisma.lab.findFirst({
                     where: {
                         name: name,
                         semester: semester,
-                        department: user.role == 'admin' && department ? department : user.department,
-                        organisation: user.organisation,
+                        department: user.role == "admin" && department ? department : user.department,
+                        orgId: user.orgId,
                     },
                 });
                 if (existingLab) {
@@ -73,16 +68,19 @@ function createLab(JWTtoken, name, semester, batches, teachers, rooms, timetable
                         name: name,
                         semester: semester,
                         batches: batches.join(";"),
-                        teachers: teachers.map(batch => batch.join(",")).join(";"),
-                        rooms: rooms.map(batch => batch.join(",")).join(";"),
-                        timetable: timetables.map(row => row.join(",")).join(";"),
-                        department: user.role == 'admin' && department ? department : user.department,
-                        organisation: user.organisation,
+                        teachers: teachers.map((batch) => batch.join(",")).join(";"),
+                        rooms: rooms.map((batch) => batch.join(",")).join(";"),
+                        timetable: timetables.map((row) => row.join(",")).join(";"),
+                        department: user.role == "admin" && department ? department : user.department,
+                        orgId: user.orgId,
                     },
                 });
                 return { status: statusCodes_1.statusCodes.CREATED, data: lab };
             }
-            return { status: status == statusCodes_1.statusCodes.OK ? statusCodes_1.statusCodes.FORBIDDEN : status, data: null };
+            return {
+                status: status == statusCodes_1.statusCodes.OK ? statusCodes_1.statusCodes.FORBIDDEN : status,
+                data: null,
+            };
         }
         catch (error) {
             console.error(error);
@@ -95,20 +93,25 @@ function deleteLabs(JWTtoken, labs) {
     return __awaiter(this, void 0, void 0, function* () {
         try {
             const { status, user } = yield auth.getPosition(JWTtoken);
+            if ((user === null || user === void 0 ? void 0 : user.orgId) == null) {
+                return { status: statusCodes_1.statusCodes.BAD_REQUEST };
+            }
             if (status == statusCodes_1.statusCodes.OK && user && user.role != "viewer") {
                 yield prisma.lab.deleteMany({
                     where: {
-                        OR: labs.map(lab => ({
+                        OR: labs.map((lab) => ({
                             name: lab.name,
                             semester: lab.semester,
-                            department: user.role == 'admin' ? lab.department : user.department,
-                            organisation: user.organisation,
-                        }))
-                    }
+                            department: user.role == "admin" ? lab.department : user.department,
+                            orgId: user.orgId,
+                        })),
+                    },
                 });
                 return { status: statusCodes_1.statusCodes.OK };
             }
-            return { status: status == statusCodes_1.statusCodes.OK ? statusCodes_1.statusCodes.UNAUTHORIZED : status };
+            return {
+                status: status == statusCodes_1.statusCodes.OK ? statusCodes_1.statusCodes.UNAUTHORIZED : status,
+            };
         }
         catch (error) {
             console.error(error);
@@ -120,13 +123,19 @@ function getLabs(JWTtoken_1) {
     return __awaiter(this, arguments, void 0, function* (JWTtoken, department = null, semester = null) {
         try {
             const { status, user } = yield auth.getPosition(JWTtoken);
+            if ((user === null || user === void 0 ? void 0 : user.orgId) == null) {
+                return {
+                    status: statusCodes_1.statusCodes.BAD_REQUEST,
+                    data: null,
+                };
+            }
             if (status == statusCodes_1.statusCodes.OK && user) {
                 const labs = yield prisma.lab.findMany({
                     where: {
-                        organisation: user.organisation,
-                        department: user.role == 'admin' && department ? department : user.department,
-                        semester: semester
-                    }
+                        orgId: user.orgId,
+                        department: user.role == "admin" && department ? department : user.department,
+                        semester: semester,
+                    },
                 });
                 return { status: statusCodes_1.statusCodes.OK, data: labs };
             }
@@ -142,13 +151,21 @@ function updateLab(JWTtoken, originalName, originalSemester, lab, originalDepart
     return __awaiter(this, void 0, void 0, function* () {
         try {
             const { status, user } = yield auth.getPosition(JWTtoken);
+            if ((user === null || user === void 0 ? void 0 : user.orgId) == null) {
+                return {
+                    status: statusCodes_1.statusCodes.BAD_REQUEST,
+                    data: null,
+                };
+            }
             if (status == statusCodes_1.statusCodes.OK && user && user.role != "viewer") {
                 const existingLab = yield prisma.lab.findFirst({
                     where: {
                         name: originalName,
                         semester: originalSemester,
-                        department: user.role == 'admin' && originalDepartment ? originalDepartment : user.department,
-                        organisation: user.organisation,
+                        department: user.role == "admin" && originalDepartment
+                            ? originalDepartment
+                            : user.department,
+                        orgId: user.orgId,
                     },
                 });
                 if (!existingLab) {
@@ -163,12 +180,17 @@ function updateLab(JWTtoken, originalName, originalSemester, lab, originalDepart
                         teachers: lab.teachers,
                         rooms: lab.rooms,
                         timetable: lab.timetable,
-                        department: user.role == 'admin' && lab.department ? lab.department : user.department,
+                        department: user.role == "admin" && lab.department
+                            ? lab.department
+                            : user.department,
                     },
                 });
                 return { status: statusCodes_1.statusCodes.OK, data: updatedLab };
             }
-            return { status: status == statusCodes_1.statusCodes.OK ? statusCodes_1.statusCodes.FORBIDDEN : status, data: null };
+            return {
+                status: status == statusCodes_1.statusCodes.OK ? statusCodes_1.statusCodes.FORBIDDEN : status,
+                data: null,
+            };
         }
         catch (error) {
             console.error(error);
@@ -180,16 +202,25 @@ function peekLab(JWTtoken, name, semester, department) {
     return __awaiter(this, void 0, void 0, function* () {
         try {
             const { status, user } = yield auth.getPosition(JWTtoken);
+            if ((user === null || user === void 0 ? void 0 : user.orgId) == null) {
+                return {
+                    status: statusCodes_1.statusCodes.BAD_REQUEST,
+                    data: null,
+                };
+            }
             if (status == statusCodes_1.statusCodes.OK && user) {
                 const lab = yield prisma.lab.findFirst({
                     where: {
                         name: name,
                         semester: semester,
-                        department: user.role == 'admin' && department ? department : user.department,
-                        organisation: user.organisation,
+                        department: user.role == "admin" && department ? department : user.department,
+                        orgId: user.orgId,
                     },
                 });
-                return { status: lab ? statusCodes_1.statusCodes.OK : statusCodes_1.statusCodes.NOT_FOUND, data: lab };
+                return {
+                    status: lab ? statusCodes_1.statusCodes.OK : statusCodes_1.statusCodes.NOT_FOUND,
+                    data: lab,
+                };
             }
             return { status: status, data: null };
         }
