+++ backend/lib/actions/teacher.js
@@ -1,5 +1,4 @@
 "use strict";
-"use server";
 var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
     if (k2 === undefined) k2 = k;
     var desc = Object.getOwnPropertyDescriptor(m, k);
@@ -16,23 +15,13 @@ var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (
 }) : function(o, v) {
     o["default"] = v;
 });
-var __importStar = (this && this.__importStar) || (function () {
-    var ownKeys = function(o) {
-        ownKeys = Object.getOwnPropertyNames || function (o) {
-            var ar = [];
-            for (var k in o) if (Object.prototype.hasOwnProperty.call(o, k)) ar[ar.length] = k;
-            return ar;
-        };
-        return ownKeys(o);
-    };
-    return function (mod) {
-        if (mod && mod.__esModule) return mod;
-        var result = {};
-        if (mod != null) for (var k = ownKeys(mod), i = 0; i < k.length; i++) if (k[i] !== "default") __createBinding(result, mod, k[i]);
-        __setModuleDefault(result, mod);
-        return result;
-    };
-})();
+var __importStar = (this && this.__importStar) || function (mod) {
+    if (mod && mod.__esModule) return mod;
+    var result = {};
+    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
+    __setModuleDefault(result, mod);
+    return result;
+};
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
     function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
@@ -54,28 +43,34 @@ const client_1 = require("@prisma/client");
 const statusCodes_1 = require("../types/statusCodes");
 const prisma = new client_1.PrismaClient();
 function convertTableToString(timetable) {
-    return timetable.map(row => row.join(",")).join(";");
+    return timetable.map((row) => row.join(",")).join(";");
 }
 function createTeachers(JWTtoken_1, name_1) {
     return __awaiter(this, arguments, void 0, function* (JWTtoken, name, initials = null, email = null, department = null, alternateDepartments = null, timetable = null, labtable = null) {
         try {
             const { status, user } = yield auth.getPosition(JWTtoken);
+            if ((user === null || user === void 0 ? void 0 : user.orgId) == null) {
+                return {
+                    status: statusCodes_1.statusCodes.BAD_REQUEST,
+                    teacher: null,
+                };
+            }
             if (status == statusCodes_1.statusCodes.OK && user) {
                 //if user isnt a viewer
-                if (user.role != 'viewer') {
+                if (user.role != "viewer") {
                     //check if teacher with same name dep and org exist
                     const teachers = yield prisma.teacher.findFirst({
                         where: {
                             name: name,
                             department: department ? department : user.department,
-                            organisation: user.organisation
-                        }
+                            orgId: user.orgId,
+                        },
                     });
                     //if even a single teacher exists
                     if (teachers) {
                         return {
                             status: statusCodes_1.statusCodes.BAD_REQUEST,
-                            teacher: null
+                            teacher: null,
                         };
                     }
                     //else
@@ -83,36 +78,44 @@ function createTeachers(JWTtoken_1, name_1) {
                         name: name,
                         initials: initials,
                         email: email,
-                        department: department ? department : user.department ? user.department : "no department",
+                        department: department
+                            ? department
+                            : user.department
+                                ? user.department
+                                : "no department",
                         alternateDepartments: alternateDepartments,
-                        timetable: timetable ? convertTableToString(timetable) : "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
-                        labtable: labtable ? convertTableToString(labtable) : "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
-                        organisation: user.organisation
+                        timetable: timetable
+                            ? convertTableToString(timetable)
+                            : "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
+                        labtable: labtable
+                            ? convertTableToString(labtable)
+                            : "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
+                        orgId: user.orgId,
                     };
                     yield prisma.teacher.create({
-                        data: teacher
+                        data: teacher,
                     });
                     return {
                         status: statusCodes_1.statusCodes.CREATED,
-                        teacher: teacher
+                        teacher: teacher,
                     };
                 }
                 //if user is a viewer code will reach here
                 return {
                     status: statusCodes_1.statusCodes.FORBIDDEN,
-                    teacher: null
+                    teacher: null,
                 };
             }
             //if status not ok
             return {
                 status: status,
-                teacher: null
+                teacher: null,
             };
         }
         catch (_a) {
             return {
                 status: statusCodes_1.statusCodes.INTERNAL_SERVER_ERROR,
-                teacher: null
+                teacher: null,
             };
         }
     });
@@ -121,54 +124,62 @@ function updateTeachers(JWTtoken_1, originalName_1) {
     return __awaiter(this, arguments, void 0, function* (JWTtoken, originalName, originalDepartment = null, teacher) {
         try {
             const { status, user } = yield auth.getPosition(JWTtoken);
+            if ((user === null || user === void 0 ? void 0 : user.orgId) == null) {
+                return {
+                    status: statusCodes_1.statusCodes.BAD_REQUEST,
+                    teacher: null,
+                };
+            }
             if (status == statusCodes_1.statusCodes.OK && user) {
-                if (user.role != 'viewer') {
+                if (user.role != "viewer") {
                     const teacherExists = yield prisma.teacher.findFirst({
                         where: {
                             name: originalName,
-                            department: user.role == 'admin' ? originalDepartment : user.department,
-                            organisation: user.organisation
-                        }
+                            department: user.role == "admin" ? originalDepartment : user.department,
+                            orgId: user.orgId,
+                        },
                     });
                     if (!teacherExists) {
                         return {
                             status: statusCodes_1.statusCodes.BAD_REQUEST,
-                            teacher: null
+                            teacher: null,
                         };
                     }
                     const updatedTeacher = yield prisma.teacher.update({
                         where: {
-                            id: teacherExists.id
+                            id: teacherExists.id,
                         },
                         data: {
                             name: teacher.name,
                             initials: teacher.initials,
                             email: teacher.email,
-                            department: user.role == 'admin' && teacher.department ? teacher.department : user.department,
+                            department: user.role == "admin" && teacher.department
+                                ? teacher.department
+                                : user.department,
                             alternateDepartments: teacher.alternateDepartments,
                             timetable: teacher.timetable,
                             labtable: teacher.labtable,
-                        }
+                        },
                     });
                     return {
                         status: statusCodes_1.statusCodes.OK,
-                        teacher: updatedTeacher
+                        teacher: updatedTeacher,
                     };
                 }
                 return {
                     status: statusCodes_1.statusCodes.FORBIDDEN,
-                    teacher: null
+                    teacher: null,
                 };
             }
             return {
                 status: status,
-                teacher: null
+                teacher: null,
             };
         }
         catch (_a) {
             return {
                 status: statusCodes_1.statusCodes.INTERNAL_SERVER_ERROR,
-                teacher: null
+                teacher: null,
             };
         }
     });
@@ -177,6 +188,12 @@ function createManyTeachers(JWTtoken_1, name_1, initials_1) {
     return __awaiter(this, arguments, void 0, function* (JWTtoken, name, initials, email = null, department = null) {
         try {
             const { status, user } = yield auth.getPosition(JWTtoken);
+            if ((user === null || user === void 0 ? void 0 : user.orgId) == null) {
+                return {
+                    status: statusCodes_1.statusCodes.BAD_REQUEST,
+                    teachers: null,
+                };
+            }
             if (status == statusCodes_1.statusCodes.OK && user && user.role != "viewer") {
                 const teachers = [];
                 for (let i = 0; i < name.length; i++) {
@@ -188,12 +205,12 @@ function createManyTeachers(JWTtoken_1, name_1, initials_1) {
                         alternateDepartments: null,
                         timetable: "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
                         labtable: "0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;0,0,0,0,0,0;",
-                        organisation: user.organisation
+                        orgId: user.orgId,
                     });
                 }
                 const duplicateChecks = yield Promise.all(teachers.map((teacher) => prisma.teacher.findFirst({
                     where: {
-                        organisation: teacher.organisation,
+                        orgId: teacher.orgId,
                         department: teacher.department,
                         name: teacher.name,
                     },
@@ -226,63 +243,73 @@ function createManyTeachers(JWTtoken_1, name_1, initials_1) {
         }
     });
 }
-//to display teachers list 
+//to display teachers list
 function getTeachers(JWTtoken) {
     return __awaiter(this, void 0, void 0, function* () {
         try {
             const { status, user } = yield auth.getPosition(JWTtoken);
+            if ((user === null || user === void 0 ? void 0 : user.orgId) == null) {
+                return {
+                    status: statusCodes_1.statusCodes.BAD_REQUEST,
+                    teachers: null,
+                };
+            }
             //if verification of roles is ok
             if (status == statusCodes_1.statusCodes.OK && user) {
                 let teachers;
                 //if role isnt admin, return teachers from same department
-                if (user.role != 'admin') {
-                    teachers = yield prisma.teacher.findMany({
+                if (user.role != "admin") {
+                    teachers = yield prisma.teacher
+                        .findMany({
                         where: {
-                            organisation: user.organisation,
-                            department: user.department
+                            orgId: user.orgId,
+                            department: user.department,
                         },
                         select: {
                             name: true,
                             department: true,
                             initials: true,
                             email: true,
-                        }
+                            orgId: true,
+                        },
                     })
                         //convert the returned object into Teacher[] type
-                        .then((teachers) => teachers.map((teacher) => (Object.assign(Object.assign({}, teacher), { alternateDepartments: null, organisation: user.organisation, timetable: null, labtable: null }))));
+                        .then((teachers) => teachers.map((teacher) => (Object.assign(Object.assign({}, teacher), { alternateDepartments: null, timetable: null, labtable: null }))));
                 }
                 //if role is admin, return teachers from all departments
                 else {
-                    teachers = yield prisma.teacher.findMany({
+                    teachers = yield prisma.teacher
+                        .findMany({
                         where: {
-                            organisation: user.organisation,
+                            orgId: user.orgId,
                         },
                         select: {
                             name: true,
                             department: true,
                             initials: true,
-                            email: true
-                        }
+                            email: true,
+                            orgId: true,
+                        },
                     })
                         //convert the returned object into Teacher[] type
-                        .then((teachers) => teachers.map((teacher) => (Object.assign(Object.assign({}, teacher), { alternateDepartments: null, organisation: user.organisation, timetable: null, labtable: null }))));
+                        .then((teachers) => teachers.map((teacher) => (Object.assign(Object.assign({}, teacher), { alternateDepartments: null, timetable: null, labtable: null }))));
                 }
                 return {
                     status: statusCodes_1.statusCodes.OK,
-                    teachers: teachers
+                    teachers: teachers,
                 };
             }
             else {
                 return {
                     status: status,
-                    teachers: null
+                    teachers: null,
                 };
             }
         }
         catch (_a) {
             return {
                 status: statusCodes_1.statusCodes.INTERNAL_SERVER_ERROR,
-                teachers: null
+                teachers: null,
             };
         }
     });
@@ -292,17 +319,27 @@ function peekTeacher(token_1, name_1) {
         try {
             //get position of user
             const { status, user } = yield auth.getPosition(token);
+            if ((user === null || user === void 0 ? void 0 : user.orgId) == null) {
+                return {
+                    status: statusCodes_1.statusCodes.BAD_REQUEST,
+                    teacher: null,
+                };
+            }
             //if verification of rules is okay, perform the following
             if (status == statusCodes_1.statusCodes.OK && user) {
                 const teacher = yield prisma.teacher.findFirst({
                     where: {
                         name: name,
-                        department: user.role == 'admin' ? department ? department : user.department : user.department,
-                        organisation: user.organisation,
+                        department: user.role == "admin"
+                            ? department
+                                ? department
+                                : user.department
+                            : user.department,
+                        orgId: user.orgId,
                     },
                     select: {
                         name: true,
-                        organisation: true,
+                        orgId: true,
                         department: true,
                         alternateDepartments: true,
                         initials: true,
@@ -316,10 +353,10 @@ function peekTeacher(token_1, name_1) {
                     teacher: teacher,
                 };
             }
-            //else 
+            //else
             return {
                 status: status,
-                teacher: null
+                teacher: null,
             };
         }
         catch (_a) {
@@ -334,35 +371,40 @@ function peekTeacher(token_1, name_1) {
 function deleteTeachers(JWTtoken, teachers) {
     return __awaiter(this, void 0, void 0, function* () {
         const { status, user } = yield auth.getPosition(JWTtoken);
+        if ((user === null || user === void 0 ? void 0 : user.orgId) == null) {
+            return {
+                status: statusCodes_1.statusCodes.BAD_REQUEST,
+            };
+        }
         try {
             if (status == 200 && user) {
-                if (user.role != 'viewer') {
+                if (user.role != "viewer") {
                     yield prisma.teacher.deleteMany({
                         where: {
-                            OR: teachers.map(teacher => ({
+                            OR: teachers.map((teacher) => ({
                                 name: teacher.name,
-                                organisation: user.organisation,
-                                department: user.role == 'admin' ? teacher.department : user.department
-                            }))
-                        }
+                                orgId: user.orgId,
+                                department: user.role == "admin" ? teacher.department : user.department,
+                            })),
+                        },
                     });
                     return {
-                        status: statusCodes_1.statusCodes.OK
+                        status: statusCodes_1.statusCodes.OK,
                     };
                 }
                 //else
                 return {
-                    status: statusCodes_1.statusCodes.FORBIDDEN
+                    status: statusCodes_1.statusCodes.FORBIDDEN,
                 };
             }
             //else
             return {
-                status: status
+                status: status,
             };
         }
         catch (_a) {
             return {
-                status: statusCodes_1.statusCodes.INTERNAL_SERVER_ERROR
+                status: statusCodes_1.statusCodes.INTERNAL_SERVER_ERROR,
             };
         }
     });
